requirements.txt
click==8.0.4
Flask==2.0.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
numpy==1.22.3
pandas==1.4.1
prettytable==3.2.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
six==1.16.0
wcwidth==0.2.5
Werkzeug==2.0.3
